+++ FRS/src/main/java/com/group/FRS/repository/RouteRepository.java
@@ -7,6 +7,6 @@ import com.group.FRS.model.Route;
 
 
 @Repository
-public interface RouteRepository extends JpaRepository<Route, Long> {
+public interface RouteRepository extends JpaRepository<Route, Integer> {
 	//
 	}
